# spec/02-engine/openapi_v1.yaml
openapi: 3.0.3
info:
  title: 등기 대행 플랫폼 API (v1)
  version: v1
  description: |
    본 OpenAPI는 `spec/02-engine/03-api-and-events.md`의 REST 예시를 “계약 파일”로 내린 버전입니다.
    - UI 렌더링은 카드(JSON) 기반이며, 카드의 엄격한 계약은 `02-engine/contracts/ui_cards.schema.json`을 우선합니다.
    - 결제/정산/리스크 등은 이벤트 저장소(`domain_events`)와 함께 동작합니다.
servers:
  - url: /
...
paths:
...
  /v1/cases:
    post:
      tags: [Cases]
      summary: 케이스 생성(파트너 선택 확정)
      operationId: postCases
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreateRequest'
      responses:
        '200':
          description: 케이스 생성 결과 + 초기 워크스페이스 카드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/cases/{caseId}:
    get:
      tags: [Cases]
      summary: 케이스 조회(요약+카드+타임라인)
      operationId: getCaseById
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 케이스 상태/체크리스트/문서슬롯/타임라인/요약카드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseGetResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: { type: string }
      description: |
        요청 추적용 ID. 클라이언트가 넣지 않으면 서버가 발급하며, 이벤트의 trace.requestId에 저장한다.
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: { type: string }
      description: |
        여러 요청/이벤트를 묶는 상위 상관관계 ID(선택).
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: |
        멱등성 키. POST 계열(특히 결제/문서/케이스 생성)은 키 사용을 강권한다.
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 충돌(상태 불일치/중복 처리 등)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: 레이트리밋 초과
      headers:
        Retry-After:
          schema: { type: string }
          description: 재시도까지 대기 시간(초 또는 HTTP-date)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: 서버 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [code, messageKo, requestId]
      properties:
        code:
          type: string
          description: "에러 코드(문서: 02-engine/06-api-error-and-idempotency.md)"
        messageKo:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    CardsResponse:
      type: object
      description: |
        카드 응답의 OpenAPI 최소 스키마. 실제 검증은 `ui_cards.schema.json`을 사용한다.
      required: [version, cards]
      properties:
        version: { type: string, example: v1 }
        cards:
          type: array
          items:
            type: object
            additionalProperties: true
...
    CaseCreateRequest:
      type: object
      required: [sessionId, selectedPartnerId, finalCasePackId, caseInput]
      properties:
        sessionId: { type: string }
        selectedPartnerId: { type: string }
        finalCasePackId: { type: string }
        caseInput:
          type: object
          additionalProperties: true
    CaseCreateResponse:
      type: object
      required: [caseId, cards]
      properties:
        caseId: { type: string }
        cards:
          $ref: '#/components/schemas/CardsResponse'
    CaseGetResponse:
      type: object
      required: [caseId, status, cards]
      properties:
        caseId: { type: string }
        status: { type: string }
        cards: { $ref: '#/components/schemas/CardsResponse' }
        timeline:
          type: array
          description: 케이스 이벤트 타임라인(요약). 원문은 별도 endpoint로 확장 가능.
          items:
            type: object
            additionalProperties: true
...